# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "药房进销存与处方调配后端服务",
        "title": "药房管理系统 API",
        "contact": {},
        "version": "1.3.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/drug-suppliers/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "删除供货关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "关系ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
        "/drugs/{id}/suppliers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "药品供货关系列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "药品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "绑定药品-供应商关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "药品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "关系信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.bindDrugSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
        "/suppliers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "供应商列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "新增供应商",
                "parameters": [
                    {
                        "description": "供应商",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.Supplier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/suppliers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "供应商详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "供应商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "更新供应商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "供应商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "供应商",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.Supplier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "删除供应商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "供应商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handler.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.bindDrugSupplierRequest": {
            "type": "object",
            "required": [
                "supplier_id"
            ],
            "properties": {
                "is_default": {
                    "type": "boolean"
                },
                "purchase_price": {
                    "type": "integer"
                },
                "supplier_id": {
                    "type": "integer"
                }
            }
        },
...
        "yaofang_internal_model.Supplier": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "contact_person": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "qualification_no": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}